fix a gcc warning
git-svn-id: file:///usr/local/Singular/svn/trunk@12911 2c84dea3-7e68-4137-9b89-c4e89433aadc

+++ Singular/iparith.cc
@@ -1078,7 +1078,7 @@ static BOOLEAN jjTIMES_P(leftv res, leftv u, leftv v)
   a=pCopy((poly)u->Data());
   b=(poly)v->CopyD(POLY_CMD); // works also for VECTOR_CMD
   if ((a!=NULL) && (b!=NULL)
-  && (pTotaldegree(a)+pTotaldegree(b)>=currRing->bitmask))
+  && ((unsigned long)(pTotaldegree(a)+pTotaldegree(b))>=currRing->bitmask))
   {
     pDelete(&a);
     pDelete(&b);
